refactor - add entities into packages
